requirements.txt
numpy==1.24.4
scipy==1.15.2
torch==2.7.0+cpu
torchvision==0.22.0+cpu
paddleocr==2.10.0
torchaudio==2.7.0+cpu
opencv-python==4.11.0
elasticsearch==8.11.1
PyQt6==6.4.2
PyQt6-Qt6==6.4.2
loguru==0.7.3
transformers==4.21.1
pandas==2.2.3
tqdm==4.67.1
requests==2.32.3
sentence_transformers==2.2.2
pillow==11.0.0
pytesseract==0.3.13
easyocr==1.7.2
"""
# Package                  Version
# ------------------------ -----------
# aiohappyeyeballs         2.6.1
# aiohttp                  3.11.18
# aiosignal                1.3.2
# albucore                 0.0.24
# albumentations           2.0.6
# annotated-types          0.7.0
# antlr4-python3-runtime   4.9.3
# anyio                    4.9.0
# astor                    0.8.1
# async-timeout            5.0.1
# attrs                    25.3.0
# beautifulsoup4           4.13.4
# certifi                  2025.4.26
# charset-normalizer       3.4.2
# click                    8.1.8
# colorama                 0.4.6
# coloredlogs              15.0.1
# colorlog                 6.9.0
# contourpy                1.2.1
# cycler                   0.12.1
# Cython                   3.0.12
# decorator                5.2.1
# easyocr                  1.7.2
# elastic-transport        8.17.1
# elasticsearch            8.11.1
# exceptiongroup           1.2.2
# filelock                 3.13.1
# fire                     0.7.0
# flatbuffers              25.2.10
# fonttools                4.53.1
# frozenlist               1.6.0
# fsspec                   2024.6.1
# h11                      0.16.0
# httpcore                 1.0.9
# httpx                    0.28.1
# huggingface-hub          0.31.1
# humanfriendly            10.0
# idna                     3.10
# imageio                  2.37.0
# Jinja2                   3.1.3
# joblib                   1.5.0
# kiwisolver               1.4.5
# lazy_loader              0.4
# lmdb                     1.6.2
# loguru                   0.7.3
# lxml                     5.4.0
# MarkupSafe               2.1.5
# matplotlib               3.9.1
# mpmath                   1.3.0
# multidict                6.4.3
# networkx                 3.3
# ninja                    1.11.1.4
# nltk                     3.9.1
# numpy                    1.24.4
# nvidia-cublas-cu12       12.6.4.1
# nvidia-cuda-runtime-cu12 12.6.77
# nvidia-cudnn-cu12        9.5.1.17
# nvidia-cufft-cu12        11.3.0.4
# nvidia-curand-cu12       10.3.7.77
# nvidia-cusolver-cu12     11.7.1.2
# nvidia-cusparse-cu12     12.5.4.2
# nvidia-nvjitlink-cu12    12.8.93
# omegaconf                2.3.0
# onnxruntime              1.16.3
# opencv-contrib-python    4.11.0.86
# opencv-python            4.11.0.86
# opencv-python-headless   4.11.0.86
# opt-einsum               3.3.0
# packaging                24.1
# paddleocr                2.10.0
# paddlepaddle             3.0.0
# paddlepaddle-gpu         3.0.0
# pandas                   2.2.3
# pillow                   11.0.0
# pip                      25.1.1
# propcache                0.3.1
# protobuf                 6.30.2
# psutil                   7.0.0
# pyclipper                1.3.0.post6
# pydantic                 2.11.4
# pydantic_core            2.33.2
# pyparsing                3.1.2
# PyQt6                    6.9.0
# PyQt6-Qt6                6.9.0
# PyQt6_sip                13.10.0
# pyreadline3              3.5.4
# pytesseract              0.3.13
# python-bidi              0.6.6
# python-dateutil          2.9.0.post0
# python-docx              1.1.2
# pytz                     2025.2
# PyYAML                   6.0.2
# RapidFuzz                3.13.0
# rapidocr-onnxruntime     1.4.4
# regex                    2024.11.6
# requests                 2.32.3
# safetensors              0.5.3
# scikit-image             0.25.2
# scikit-learn             1.6.1
# scipy                    1.15.2
# sentence-transformers    4.1.0
# sentencepiece            0.2.0
# setuptools               67.6.0
# shapely                  2.1.0
# simsimd                  6.2.1
# six                      1.16.0
# sniffio                  1.3.1
# soupsieve                2.7
# stringzilla              3.12.5
# sympy                    1.13.3
# termcolor                3.1.0
# tesseract                0.1.3
# threadpoolctl            3.6.0
# tifffile                 2025.3.30
# tokenizers               0.21.1
# torch                    2.7.0+cpu
# torchaudio               2.7.0+cpu
# torchvision              0.22.0+cpu
# tqdm                     4.67.1
# transformers             4.51.3
# typing_extensions        4.13.2
# typing-inspection        0.4.0
# tzdata                   2025.2
# urllib3                  2.4.0
# win32_setctime           1.2.0
# yarl                     1.20.0
"""